Modularized the datatag angular module
